+++ test/test_trainer.py
@@ -16,7 +16,7 @@ from torch import nn
 
 try:
     from tensorboard.backend.event_processing import event_accumulator
-    from torchrl.trainers.loggers.tensorboard import TensorboardLogger
+    from torchrl.record.loggers import TensorboardLogger
 
     _has_tb = True
 except ImportError:
